update help text and give error on folder info

+++ Misc/Copy-FileToNavContainer.ps1
@@ -1,4 +1,18 @@
-﻿function Copy-FileToNavContainer {
+﻿<# 
+ .Synopsis
+  Copy File to a container
+ .Description
+  Copies a file to a Container
+ .Parameter containerName
+  Name of the container to which you want to copy a file
+ .Parameter localPath
+  Path to the file on the host, which has to be copied
+ .Parameter containerPath
+  Path of the file in the Container. This cannot be a foldername, it needs to be a filename.
+ .Example
+  Copy-FileToNavContainer -containerName test2 -localPath "c:\temp\myfile.txt" -containerPath "c:\run\my\myfile.txt"
+#>
+function Copy-FileToNavContainer {
 
     Param(
         [Parameter(Mandatory=$true)]
@@ -25,6 +39,9 @@
             try {
                 Copy-Item -Path $localPath -Destination $tempFile
                 Invoke-ScriptInNavContainer -containerName $containerName -scriptblock { Param($tempFile, $containerPath)
+                    if (Test-Path $containerPath -PathType Container) {
+                        throw "ContainerPath ($containerPath) already exists as a folder. Cannot copy file, ContainerPath needs to specify a filename."
+                    }
                     Move-Item -Path $tempFile -Destination $containerPath -Force
                 } -argumentList $tempFile, $containerPath
             } finally {
